pyproject.toml
[tool.poetry]
name = "address-book-api"
version = "1.0"
description = ""
authors = ["Julian Tan <julian.tanz@yahoo.com>"]

[tool.poetry.dependencies]
python = "^3.9"
fastapi = "^0.68.1"
uvicorn = "^0.15.0"
python-multipart = "^0.0.5"
pyhumps = "^3.0.2"
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
python-jose = "^3.3.0"
SQLAlchemy = "^1.4.23"
requests = "^2.28.1"

[tool.poetry.dev-dependencies]
pytest = "^6.2.4"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
